# swagger_output.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HouseForSaleSearcher REST API",
    "description": "This is a OpenAPI interactive documentation for HouseForSaleSearcher's API CodersCamp project.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000/api"
    }
  ],
  "security": [
    {
      "Token": []
    }
  ],
  "paths": {
...
    "/transactions": {
      "post": {
        "tags": ["transaction"],
        "summary": "Transaction process (buy house)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "CREATED",
            "content": {}
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
    "/transactions/{id}": {
      "get": {
        "tags": ["transaction"],
        "summary": "Get transaction by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {}
          }
        }
      }
    },
    "/transactions/": {
      "get": {
        "tags": ["transaction"],
        "summary": "Get all transactions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "buyerId": {
            "type": "string"
          },
          "houseId": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Token": {
        "type": "http",
        "description": "JWT Authorization header",
        "scheme": "bearer"
      }
    }
  }
}